requirements.txt
streamlit==1.36.0

pandas==2.2.2
numpy==1.26.4

scikit-learn==1.4.2

scikit-surprise==1.1.3

joblib==1.3.2

altair==5.2.0

typing_extensions>=4.5.0
